add a v2 eip resource which also supports node selectors
this will allow us to associate eips with nodes directly, and
automatically reassociate them as nodes are replaced. need to use
selectors here because unlike pods which keep a consistent name when
they are recreated, nodes have names auto-assigned by eks, so we need
another method to identify which node goes with which eip.